Use f64 for float literals
Promote floating-point literal handling from f32 to f64 in the tokenizer and AST expression kind so numeric precision matches the language's float value semantics. This keeps parsed float values consistent across compilation stages.

+++ src/compiler/tokenizer.rs
@@ -13,7 +13,7 @@ pub enum Token {
     Identifier(String),
     ConstInteger32(i32),
     ConstInteger64(i64),
-    ConstFloat(f32),
+    ConstFloat(f64),
     ConstString(Vec<PosChar>),
     ConstBoolean(bool),
     ConstChar(char),
@@ -215,7 +215,7 @@ fn string_to_token(string: &String) -> Token {
         }
     }
 
-    if let Ok(float) = string.parse::<f32>() {
+    if let Ok(float) = string.parse::<f64>() {
         return Token::ConstFloat(float);
     }
